+++ src/main/java/dev/chojo/ember/feature/notifications/entity/NotificationType.java
@@ -17,6 +17,7 @@ public enum NotificationType {
     EVENT_REGISTRATION_STATUS(NotificationParams.EventRegistrationStatus.class, "notification.eventRegistrationStatus"),
     EXCHANGE_STATUS_CHANGE(NotificationParams.ExchangeStatusChange.class, "notification.exchangeStatusChange"),
     EXCHANGE_NEW_REQUEST(NotificationParams.ExchangeNewRequest.class, "notification.exchangeNewRequest"),
+    MOVEMENT_DECLINED(NotificationParams.MovementDeclined.class, "notification.movementDeclined"),
     NEW_EVENT(NotificationParams.NewEvent.class, "notification.newEvent"),
     NEW_EVENTS_BATCH(NotificationParams.NewEventsBatch.class, "notification.newEventsBatch"),
     MEMBER_ADDED_TO_GROUP(NotificationParams.MemberAddedToGroup.class, "notification.memberAddedToGroup"),