+++ src/test/java/dev/chojo/ember/feature/inventory/service/MovementFlowServiceTest.java
@@ -166,6 +166,7 @@ class MovementFlowServiceTest extends RepositoryTestBase {
                 station.id(),
                 MovementPurpose.RETURN,
                 null,
+                null,
                 itemId,
                 inventoryId,
                 null,